backend/requirements.txt
Django==4.2
djangorestframework==3.15.0
django-cors-headers==4.0.0
pytz==2024.1
gunicorn==21.2.0
djangorestframework-simplejwt==5.2.2
setuptools>=65.0.0
